refactor(memory): split the 1700-line binary main into its four seams
main.rs held the whole daemon: transport selection, store opening, backend
construction, subcommands, and their tests, 1708 lines in one file. It splits
along the seams the code already had:

  daemon_serve     transport & lifecycle: stdio vs HTTP, the loopback gate,
                   logging, shutdown signals, the orphan watchdog
  daemon_startup   opening the store and assembling the configured server:
                   lock retries, provenance, autograph, config-file and TTL
  daemon_backends  building the configured embedder and extractor from the
                   environment, and the startup reachability warning
  daemon_commands  the early subcommands that never open the live store as a
                   daemon: export, migrate-embeddings, compile-stdin

main.rs keeps what a reader opens it for: main() and the early-command
dispatch, 92 lines. Code moved verbatim — the compiler enumerated the
cross-module seams (startup and commands import from backends; nothing
imports from serve but main), and each module's unit tests moved with their
subjects, unchanged.

Also records the env-read audit the review asked for, on config.rs where the
export-into-env rule lives: every env::var in the library sits in a module
whose purpose is resolving daemon configuration, plus the one documented
per-call knob (the embedder keep-alive, whose read-per-call timing is the
contract). A new env read anywhere else is a defect against that paragraph —
the finding is resolved by classification, not by churn that would have moved
reads out of the modules that are the doors.

Verified: all-features and default builds clean, pedantic clippy clean on all
targets, the binary's 8 unit tests and the daemon process tests (http
transport, graceful shutdown, store path) green against the split binary,
lizard shows zero functions over the gate in the new modules.
